requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
